fixed unsigned compares
git-svn-id: svn://svn.lighttpd.net/lighttpd/branches/lighttpd-merge-1.4.x@855 152afb58-edef-0310-8abb-c4023f1b3aa9

+++ src/mod_dirlisting.c
@@ -637,7 +637,7 @@ static int http_list_directory(server *srv, connection *con, plugin_data *p, buf
 		/* NOTE: the manual says, d_name is never more than NAME_MAX
 		 *       so this should actually not be a buffer-overflow-risk
 		 */
-		if (i > name_max) continue;
+		if (i > (size_t)name_max) continue;
 		
 		memcpy(path_file, dent->d_name, i + 1);
 		if (stat(path, &st) != 0)
